Implementing lexer and so called compiler
- changing uint8_t type to uint32_t (do i use too much memory?)
- 1 + 2 works, just once tho but fuck it at least it works
- i dont fucking know what's going on brokma saket utak ke

+++ src/chunk.c
@@ -11,12 +11,12 @@ void init_chunk(Chunk *chunk) {
     init_value_array(&(chunk->constants));
 }
 
-void append_chunk(Chunk *chunk, uint8_t byte, uint8_t line) {
+void append_chunk(Chunk *chunk, uint8_t byte, uint32_t line) {
     if (chunk->capacity < chunk->used_count + 1) { // capacity not enough so we extend it, also +1 because we want to add a byte
-        uint8_t old_capacity = chunk->capacity;
+        uint32_t old_capacity = chunk->capacity;
         chunk->capacity = GROW_CAPACITY(old_capacity);
-        chunk->code = (uint8_t *) dynamic_realloc(chunk->code, old_capacity, chunk->capacity);
-        chunk->lines = (uint8_t *) dynamic_realloc(chunk->lines, old_capacity, chunk->capacity);
+        chunk->code = (uint32_t *) dynamic_realloc(chunk->code, old_capacity, chunk->capacity);
+        chunk->lines = (uint32_t *) dynamic_realloc(chunk->lines, old_capacity, chunk->capacity);
     }
     chunk->code[chunk->used_count] = byte; // e.g. chunk->code[0] = byte
     chunk->lines[chunk->used_count] = line;
@@ -32,7 +32,7 @@ void free_chunk(Chunk *chunk) {
     init_chunk(chunk);
 }
 
-uint8_t add_constant(Chunk *chunk, Value value) {
+uint32_t add_constant(Chunk *chunk, Value value) {
     append_value_array(&(chunk->constants), value);
     return  chunk->constants.used_count - 1; // After adding the constant, we return the index where the constant was appended so that we can locate that same constant later.
 }
