+++ src/debug.h
@@ -3,9 +3,9 @@
 #include <stdio.h>
 #include "chunk.h"
 
-uint8_t disassemble_instruction(Chunk *chunk, uint8_t offset);
+uint32_t disassemble_instruction(Chunk *chunk, uint32_t offset);
 void disassemble_chunk(Chunk *chunk, const char *name);
-uint8_t disassemble_constant_instruction(Chunk *chunk, uint8_t offset);
+uint32_t disassemble_constant_instruction(Chunk *chunk, uint32_t offset);
 
 
 #endif // ATJEH_DEBUG_H
